gooo

+++ src/app/components/gestion/gestion.component.ts
@@ -6,6 +6,7 @@ import { DialogosService } from '../../services/dialogos.service';
 import { GestionService } from 'src/app/services/gestion.service';
 import { NavigationHeaderService } from 'src/app/services/navigation-header.service';
 import { Router } from '@angular/router';
+import { DialogTypes } from '../../components/dialogos/dialogos-general';
 
 @Component({
   selector: 'app-gestion',
@@ -19,26 +20,26 @@ export class GestionComponent implements OnInit {
 
   dataSourceTabla: MatTableDataSource<UsuMezcla>
   columnas: string[] = ['Nombre', 'Apellidos', 'Telefono', 'Edad', 'Direccion', 'Dni', 'Info'
-  , 'Nivel', 'Email', 'Actualizar', 'Eliminar'];
+    , 'Nivel', 'Email', 'Actualizar', 'Eliminar'];
 
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSourceTabla.filter = filterValue.trim().toLowerCase();
   }
-  constructor(private gestionService: GestionService, private dialogosService: DialogosService, 
-    private router: Router,private navigationHeaderService: NavigationHeaderService) { }
+  constructor(private gestionService: GestionService, private dialogosService: DialogosService,
+    private router: Router, private navigationHeaderService: NavigationHeaderService) { }
 
   ngOnInit(): void {
     this.navigationHeaderService.getDatosUsuario().subscribe(data => {
-    
+
       this.usuarioAutenticado2 = data['rol'];
-     if(this.usuarioAutenticado2 != 1){
-      this.router.navigate(['/welcome']);
-     }
+      if (this.usuarioAutenticado2 != 1) {
+        this.router.navigate(['/welcome']);
+      }
       console.log(this.usuarioAutenticado);
     });
 
-    this.gestionService.getDatosTodasLosUsuarios().subscribe(data =>{
+    this.gestionService.getDatosTodasLosUsuarios().subscribe(data => {
       this.dataSourceTabla = new MatTableDataSource<UsuMezcla>(data['todasLosDatosUsuario']);
       //también se lo asignamos a este listado para poder usarlo
       this.listadoTodosLosUsuarios = data['todasLosDatosUsuario'];
@@ -49,16 +50,17 @@ export class GestionComponent implements OnInit {
    * 
    */
 
-   deleteUser(id_usuario){
-    //this.dialogosService.abrirDialogCargando();
-    this.gestionService.deleteUsers(id_usuario).subscribe(data => {
-      //        console.log(data);
-    //this.dialogosService.cerrarDialogo();
-    this.dialogosService.abrirDialogInfo("¡Este usuario ha sido eliminado!").subscribe(opcionElegida => {
-      window.location.reload();
+  deleteUser(id_usuario) {
+
+
+
+    this.dialogosService.abrirDialogConfirmacion("¿Quieres eliminar a este usuario?").subscribe(opcionElegida => {
+      if (opcionElegida == DialogTypes.RESPUESTA_ACEPTAR) {
+        this.gestionService.deleteUsers(id_usuario).subscribe(data => {
+          window.location.reload();
         });
-      
-      });
+      }
+    });
   }
 
 }
